- refactoring вынесение кода подтверждения в отдельный класс
